+++ test/config.test.js
@@ -13,13 +13,12 @@ describe('config', function () {
       , jadeFiles = [ fixturesPath + 'test.jade', fixturesPath + 'test-include.jade' ]
       , options = { stylperjaderc: 'nonexistent' }
 
-    assert.throws(function () {
-      stylperjade(cssFiles, jadeFiles, options, function (err, results) {
-        assert.equal(results, null)
-      })
-    }
-    , /.stylperjaderc not found/)
-    done()
+    stylperjade(cssFiles, jadeFiles, options, function (err, results) {
+      assert(err)
+      assert.equal(err, '.stylperjaderc not found')
+      assert.equal(results, null)
+      done()
+    })
   })
 
   it('should error if options.stylperjaderc is invalid', function (done) {
@@ -27,13 +26,12 @@ describe('config', function () {
       , jadeFiles = [ fixturesPath + 'test.jade', fixturesPath + 'test-include.jade' ]
       , options = { stylperjaderc: fixturesPath + 'invalid.json' }
 
-    assert.throws(function () {
-      stylperjade(cssFiles, jadeFiles, options, function (err, results) {
-        assert.equal(results, null)
-      })
-    }
-    , /.stylperjaderc is invalid JSON/)
-    done()
+    stylperjade(cssFiles, jadeFiles, options, function (err, results) {
+      assert(err)
+      assert.equal(err, '.stylperjaderc is invalid JSON')
+      assert.equal(results, null)
+      done()
+    })
   })
 
   it('should load config from options.stylperjaderc', function (done) {
@@ -50,8 +48,8 @@ describe('config', function () {
       assert.equal(results.blacklistedTotal, 0)
       assert.equal(results.blacklistedCssCount, 0)
       assert.equal(results.blacklistedJadeCount, 0)
-      assert.equal(chalk.stripColor(results.report.trim())
-        , expectedReport.replace(/%dirname%/g, __dirname).trim()
+      assert.equal(chalk.stripColor(results.report)
+        , expectedReport.replace(/%dirname%/g, __dirname)
         , results.report)
       done()
     })
@@ -78,8 +76,8 @@ describe('config', function () {
       assert.equal(_.findIndex(results.unusedJadeClasses, 'name', 'epsilon') !== -1, true)
       assert.equal(_.findIndex(results.unusedJadeClasses, 'name', 'js-alpha') !== -1, true)
       assert.equal(_.findIndex(results.unusedJadeClasses, 'name', 'beta') === -1, true)
-      assert.equal(chalk.stripColor(results.report.trim())
-        , expectedReport.replace(/%dirname%/g, __dirname).trim()
+      assert.equal(chalk.stripColor(results.report)
+        , expectedReport.replace(/%dirname%/g, __dirname)
         , results.report)
       done()
     })
@@ -99,8 +97,29 @@ describe('config', function () {
       assert.equal(results.blacklistedTotal, 0)
       assert.equal(results.blacklistedCssCount, 0)
       assert.equal(results.blacklistedJadeCount, 0)
-      assert.equal(chalk.stripColor(results.report.trim())
-        , expectedReport.replace(/%dirname%/g, __dirname).trim()
+      assert.equal(chalk.stripColor(results.report)
+        , expectedReport.replace(/%dirname%/g, __dirname)
+        , results.report)
+      done()
+    })
+  })
+
+  it('should use empty config if working directory does not contain .stylperjaderc', function (done) {
+    var cssFiles = [ '**/test.css' ]
+      , jadeFiles = [ '**/test*.jade' ]
+      , options = { cwd: __dirname }
+      , expectedReport = fs.readFileSync(fixturesPath + 'expected-unused.txt', 'utf-8')
+
+    stylperjade(cssFiles, jadeFiles, options, function (err, results) {
+      assert(!err, err)
+      assert.equal(results.unusedTotal, 18)
+      assert.equal(results.unusedCssCount, 6)
+      assert.equal(results.unusedJadeCount, 12)
+      assert.equal(results.blacklistedTotal, 0)
+      assert.equal(results.blacklistedCssCount, 0)
+      assert.equal(results.blacklistedJadeCount, 0)
+      assert.equal(chalk.stripColor(results.report)
+        , expectedReport.replace(/%dirname%/g, __dirname)
         , results.report)
       done()
     })